split node-specific stuff out of realm_js_helpers.h
