+++ packages/bindgen/src/realm_js_helpers.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include "realm/binary_data.hpp"
 #include "realm/object-store/object_store.hpp"
 #include "realm/object-store/sync/sync_session.hpp"
@@ -6,13 +8,11 @@
 #include "realm/sync/protocol.hpp"
 #include "realm/util/base64.hpp"
 #include "realm/util/logger.hpp"
-#include "realm/util/to_string.hpp"
 #include <condition_variable>
 #include <exception>
 #include <iostream>
 #include <memory>
 #include <mutex>
-#include <napi.h>
 
 #include <realm/object-store/keypath_helpers.hpp>
 #include <realm/object-store/results.hpp>
@@ -36,7 +36,7 @@
 // Equivalent to std::forward<decltype(x)>(x), but faster to compile and less impact on debug builds
 #define FWD(x) ((decltype(x)&&)(x))
 
-namespace realm::js::node {
+namespace realm::js {
 namespace {
 
 // These types are exposed to JS in the spec.
@@ -241,6 +241,7 @@ struct ObjectChangeSet {
 ////////////////////////////////////////////////////////////
 
 // These helpers are used by the generated code.
+
 template <typename Container>
 class [[nodiscard]] ContainerResizer {
 public:
@@ -265,71 +266,6 @@ private:
     Container* const m_container;
     const size_t m_old_size;
 };
-
-// TODO consider allowing Number (double) with (u)int64_t.
-inline int64_t extractInt64FromNode(const Napi::Value& input)
-{
-    bool lossless;
-    auto output = input.As<Napi::BigInt>().Int64Value(&lossless);
-    if (!lossless)
-        throw Napi::RangeError::New(input.Env(), "Value doesn't fit in int64_t");
-    return output;
-}
-inline uint64_t extractUint64FromNode(const Napi::Value& input)
-{
-    bool lossless;
-    auto output = input.As<Napi::BigInt>().Uint64Value(&lossless);
-    if (!lossless)
-        throw Napi::RangeError::New(input.Env(), "Value doesn't fit in uint64_t");
-    return output;
-}
-
-template <typename... Args>
-inline Napi::Function bindFunc(Napi::Function func, Napi::Object self, Args... args)
-{
-    return func.Get("bind").As<Napi::Function>().Call(func, {self, args...}).template As<Napi::Function>();
-}
-
-REALM_NOINLINE inline Napi::Object toNodeErrorCode(Napi::Env& env, const std::error_code& e) noexcept
-{
-    REALM_ASSERT_RELEASE(e);
-    auto out = Napi::Error::New(env, e.message()).Value();
-    out.Set("code", e.value());
-    out.Set("category", e.category().name());
-    return out;
-}
-
-REALM_NOINLINE inline Napi::Object toNodeException(Napi::Env& env, const std::exception_ptr& e) noexcept
-{
-    try {
-        std::rethrow_exception(e);
-    }
-    catch (const Napi::Error& e) {
-        return e.Value();
-    }
-    catch (const std::exception& e) {
-        return Napi::Error::New(env, e.what()).Value();
-    }
-    catch (...) {
-        return Napi::Error::New(env, "Unknown Error").Value();
-    }
-}
-
-[[noreturn]] REALM_NOINLINE inline void throwNodeException(Napi::Env& env, const std::exception& e)
-{
-    if (dynamic_cast<const Napi::Error*>(&e))
-        throw; // Just allow exception propagation to continue
-    // TODO consider throwing more specific errors in some cases.
-    // TODO consider using ThrowAsJavaScriptException instead here.
-    throw Napi::Error::New(env, e.what());
-}
-
-[[noreturn]] REALM_NOINLINE inline void throwNullSharedPtrError(Napi::Env& env, const char* clsName)
-{
-    throw Napi::Error::New(env, util::format("Attempting to use an instanace of $1 holding a null shared_ptr. "
-                                             "Did you call $resetSharedPtr on it already?",
-                                             clsName));
-}
 
 template <typename F>
 auto schedulerWrapBlockingFunction(F&& f)
@@ -368,4 +304,4 @@ auto schedulerWrapBlockingFunction(F&& f)
 }
 
 } // namespace
-} // namespace realm::js::node
+} // namespace realm::js
